pyproject.toml
[project]
name = "openvla-inference"
version = "0.1.0"
description = "Python CLI for running inference on OpenVLA vision-language-action models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "numpy>=1.21.0,<2.0.0",
    "pillow>=12.1.0",
    "protobuf>=3.20.0,<5.0.0",
    "sentencepiece>=0.2.1",
    "timm>=0.9.10,<1.0.0",
    "torch>=2.2.0",
    "transformers>=4.57.3",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
